refactor: split WebRoutes.swift into focused files + add 18 integration tests
WebRoutes.swift was 1,200 lines mixing route handlers, context structs,
and helper functions. Split into four files:

- WebRoutes.swift (322 lines) — boot, index, newSetupForm, createSetup
- WebContextTypes.swift (152 lines) — 14 view-context structs
- WebRoutes+Notebook.swift (265 lines) — notebook handlers + helpers
- WebRoutes+Submission.swift (335 lines) — submission handlers + helpers

Added WebRoutesTests.swift with 18 integration tests covering:
- Index page (redirect, empty state, visibility, grades)
- Submit form and history pages
- Submission result page (pending, results, access control)
- Tier visibility (pre/post deadline, instructor vs student)
